+++ examples/pattern_processing/pattern_binning.py
@@ -23,7 +23,7 @@ print(s)
 print(s.static_background.shape)
 print(s.detector)
 
-########################################################################################
+# %%
 # Downsample by a factor of 8 while maintaining the data type (achieved by rescaling the
 # pattern intensity). Note how the :attr:`~kikuchipy.signals.EBSD.static_background` and
 # :attr:`~kikuchipy.signals.EBSD.detector` attributes are updated.
@@ -33,7 +33,7 @@ _ = hs.plot.plot_images([s, s2], axes_decor="off", tight_layout=True, label=None
 print(s2.static_background.shape)
 print(s2.detector)
 
-########################################################################################
+# %%
 # Rebin by passing the new shape (use ``(1, 1, 60, 60)`` if binning a 2D map). Note how
 # the pattern is not rescaled and the data type is cast to either ``int64`` or
 # ``float64`` depending on the initial data type.
@@ -42,7 +42,7 @@ s3 = s.rebin(new_shape=(60, 60))
 print(s3.data.dtype)
 print(s3.data.min(), s3.data.max())
 
-########################################################################################
+# %%
 # The latter method is more flexible in that it allows for different binning factors in
 # each axis, the factors are not restricted to being integers and the factors do not
 # have to be divisors of the initial signal shape.